Question 1's answer
Q1: Write a program that will define a function which will get an array and the (5)
number to be searched. If the number is found in the array, then the function
will print the message “Search successful” and the total count of numbers,
else message “Item not found” will be printed.